+++ testing/tests/DevExpress.ui.widgets.dataGrid/dataGrid.tests.js
@@ -47,7 +47,6 @@ import '../../../node_modules/jsrender/jsrender.min.js';
 import DataGrid from 'ui/data_grid/ui.data_grid';
 import $ from 'jquery';
 import Class from 'core/class';
-import ODataUtils from 'data/odata/utils';
 import resizeCallbacks from 'core/utils/resize_callbacks';
 import { logger } from 'core/utils/console';
 import commonUtils from 'core/utils/common';
@@ -56,7 +55,6 @@ import devices from 'core/devices';
 import browser from 'core/utils/browser';
 import version from 'core/version';
 import gridCore from 'ui/data_grid/ui.data_grid.core';
-import gridCoreUtils from 'ui/grid_core/ui.grid_core.utils';
 import { DataSource } from 'data/data_source/data_source';
 import ArrayStore from 'data/array_store';
 import messageLocalization from 'localization/message';
@@ -125,41 +123,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
 
     QUnit.test('formatValue for grouped column with calculateGroupValue', function(assert) {
         assert.strictEqual(gridCore.formatValue('2012', { format: 'shortDate' }), '2012');
-    });
-
-    QUnit.test('DataGrid - Should hide filter row menu after losing it\'s focus', function(assert) {
-    // arrange
-        const filterRowWrapper = dataGridWrapper.filterRow;
-
-        createDataGrid({
-            filterRow: { visible: true },
-            dataSource: [{ field1: '1', field2: '2' }]
-        });
-        this.clock.tick();
-
-        // act
-        const $menu = filterRowWrapper.getMenuElement(0);
-        $menu.focus();
-
-        const menuInstance = $menu.dxMenu('instance');
-        const $root = $(menuInstance.itemElements().get(0));
-        menuInstance._showSubmenu($root);
-        const subMenu = menuInstance._visibleSubmenu;
-
-        // assert
-        assert.ok(subMenu._isVisible(), 'submenu exists');
-
-        // act
-        if(browser.msie && browser.version <= 11) {
-            const event = document.createEvent('Event');
-            event.initEvent('blur', true, true);
-            $menu[0].dispatchEvent(event);
-        } else {
-            $menu.blur();
-        }
-
-        // assert
-        assert.notOk(subMenu._isVisible(), 'submenu is hidden');
     });
 
     QUnit.test('commonColumnOptions', function(assert) {
@@ -4299,36 +4262,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.equal($footerView.scrollLeft(), 400, 'scroll left of footer view');
     });
 
-    // T702241
-    QUnit.test('Scroll position headers after changing of headerFilter setting', function(assert) {
-    // arrange
-        const $dataGrid = $('#dataGrid').dxDataGrid({
-            width: 200,
-            scrolling: {
-                useNative: false
-            },
-            columns: [
-                { dataField: 'firstName', width: 200 },
-                { dataField: 'lastName', width: 200 }
-            ]
-        });
-        const dataGrid = $dataGrid.dxDataGrid('instance');
-        let $headersView;
-
-        // act
-        this.clock.tick();
-
-        $headersView = $dataGrid.find('.dx-datagrid-headers' + ' .dx-datagrid-scroll-container').first();
-        $headersView.scrollLeft(200);
-        $($headersView).trigger('scroll');
-
-        dataGrid.option('headerFilter.visible', true);
-
-        // assert
-        $headersView = $dataGrid.find('.dx-datagrid-headers' + ' .dx-datagrid-scroll-container').first();
-        assert.equal($headersView.scrollLeft(), 200);
-    });
-
     QUnit.test('Horizontal scroll position of footer view is changed_T251448', function(assert) {
     // arrange
         const $dataGrid = $('#dataGrid').dxDataGrid({
@@ -4557,43 +4490,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         });
 
         assert.equal(loadCallCount, 1, 'one load count on start');
-    });
-
-    // T162735, T406593
-    QUnit.test('Load count on start when stateStoring enabled with search/filterRow values', function(assert) {
-        let loadCallCount = 0;
-        let loadFilter;
-        let contentReadyCallCount = 0;
-        const dataGrid = createDataGrid({
-            onContentReady: function() {
-                contentReadyCallCount++;
-            },
-            remoteOperations: { filtering: true, sorting: true, paging: true },
-            dataSource: {
-                load: function(options) {
-                    loadCallCount++;
-                    loadFilter = options.filter;
-                    return $.Deferred().resolve([{ field1: 'text1', field2: 100 }, { field1: 'text2', field2: 200 }], { totalCount: 2 });
-                }
-            },
-            stateStoring: {
-                enabled: true,
-                type: 'custom',
-                customLoad: function() {
-                    return {
-                        columns: [{ dataField: 'field1', dataType: 'string', visibleIndex: 0 }, { dataField: 'field2', dataType: 'number', visibleIndex: 1 }],
-                        searchText: '200'
-                    };
-                }
-            }
-
-        });
-
-        this.clock.tick();
-        assert.ok(dataGrid);
-        assert.equal(contentReadyCallCount, 1, 'contentReady is called once');
-        assert.equal(loadCallCount, 1, '1 load count on start');
-        assert.deepEqual(loadFilter, [['field1', 'contains', '200'], 'or', ['field2', '=', 200]]);
     });
 
     // T268912
@@ -4885,110 +4781,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
 
         // assert
         assert.equal(keyOfSpy.callCount, 55, 'keyOf call count');
-    });
-
-    // T802967
-    QUnit.test('calculateFilterExpression should not be called infinite times if it returns function and scrolling mode is virtual', function(assert) {
-        const data = [];
-        for(let i = 0; i < 25; i++) {
-            data.push({ test: i });
-        }
-        let calculateFilterExpressionCallCount = 0;
-        try {
-            createDataGrid({
-                loadingTimeout: undefined,
-                scrolling: {
-                    mode: 'virtual'
-                },
-                columns: [{
-                    selectedFilterOperation: '=',
-                    filterValue: [],
-                    dataField: 'test',
-                    dataType: 'number',
-                    calculateFilterExpression: function(filterValues) {
-                        calculateFilterExpressionCallCount++;
-                        return function() {
-                            return filterValues.length === 0;
-                        };
-                    }
-                }],
-                dataSource: data
-            });
-
-        } catch(err) {
-            assert.ok(false, 'the error is thrown');
-        } finally {
-            assert.equal(calculateFilterExpressionCallCount, 2, 'calculateFilterExpression call count');
-        }
-    });
-
-    // T802967
-    QUnit.test('getCombinedFilter should work correctly if filterPanel is visible and calculateFilterExpression returns function', function(assert) {
-        const data = [];
-        for(let i = 0; i < 21; i++) {
-            data.push({ test: i });
-        }
-        let calculateFilterExpressionCallCount = 0;
-        const grid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: data,
-            filterPanel: { visible: true },
-            columns: [{
-                selectedFilterOperation: '=',
-                filterValue: 0,
-                dataField: 'test',
-                calculateFilterExpression: function() {
-                    calculateFilterExpressionCallCount++;
-                    return function() {
-                        return true;
-                    };
-                }
-            }]
-        });
-
-        assert.equal(calculateFilterExpressionCallCount, 4, 'calculateFilterExpression call count');
-        assert.ok(grid.getCombinedFilter(), 'combined filter');
-        assert.equal(calculateFilterExpressionCallCount, 5, 'calculateFilterExpression call count');
-    });
-
-    // T364210
-    QUnit.test('Load count on start when EdmLiteral in calculatedFilterExpression is used and scrolling mode is virtual', function(assert) {
-        let loadCallCount = 0;
-        let contentReadyCallCount = 0;
-        const dataGrid = createDataGrid({
-            onContentReady: function() {
-                contentReadyCallCount++;
-            },
-            height: 100,
-            remoteOperations: {
-                paging: true, filtering: true
-            },
-            loadingTimeout: undefined,
-            scrolling: {
-                mode: 'virtual'
-            },
-            columns: [{
-                dataField: 'test',
-                selectedFilterOperation: '>',
-                filterValue: 50,
-                dataType: 'number',
-                calculateFilterExpression: function(value, filterOperation) {
-                    value = new ODataUtils.EdmLiteral(value + 'm');
-                    return [this.dataField, filterOperation || '=', value];
-                }
-            }],
-            dataSource: {
-                pageSize: 5,
-                load: function(options) {
-                    loadCallCount++;
-                    return $.Deferred().resolve([{}, {}, {}, {}, {}], { totalCount: 100 });
-                }
-            }
-        });
-
-        assert.ok(dataGrid);
-        assert.equal(loadCallCount, 2, 'two load count on start');
-        assert.equal(contentReadyCallCount, 1, 'one contentReady on start');
     });
 
     QUnit.test('contentReady event must be raised once when scrolling mode is virtual', function(assert) {
@@ -5502,69 +5294,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
 
         });
     }
-
-    QUnit.test('search text when scrolling mode virtual and one column is not defined', function(assert) {
-    // arrange, act
-
-        const dataSource = [
-            { 'CompanyName': 'K&S Music' },
-            { 'CompanyName': 'Super Mart of the West' },
-            { 'CompanyName': 'Electronics Depot' },
-            { 'CompanyName': 'K&S Music' },
-            { 'CompanyName': 'Kiwi Market' }
-        ];
-
-        const dataGrid = createDataGrid({
-            dataSource: dataSource,
-            loadingTimeout: undefined,
-            scrolling: {
-                mode: 'virtual'
-            },
-            searchPanel: {
-                text: 'Kiwi'
-            },
-            paging: {
-                pageSize: 2
-            },
-            columns: ['CompanyName', 'Undefined'] }
-        );
-
-        assert.equal(dataGrid.getVisibleRows().length, 1, 'items were filtered');
-    });
-
-    // T820316
-    QUnit.test('Error should not be thrown when searching text in calculated column with lookup', function(assert) {
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ text: 'text', num: 1 }, { text: 'text', num: 2 }],
-            searchPanel: {
-                visible: true
-            },
-            columns: [{
-                calculateCellValue: function(rowData) {
-                    return rowData.num;
-                },
-                allowFiltering: true,
-                lookup: {
-                    dataSource: [{ id: 1, name: 'one' }, { id: 2, name: 'two' }],
-                    valueExpr: 'id',
-                    displayExpr: 'name'
-                }
-            }, 'text']
-        });
-
-        try {
-            dataGrid.option('searchPanel.text', 'one');
-            this.clock.tick();
-        } catch(e) {
-            assert.ok(false, 'error was thrown');
-        }
-
-        const visibleRows = dataGrid.getVisibleRows();
-
-        assert.equal(visibleRows.length, 1, 'one row is visible');
-        assert.deepEqual(visibleRows[0].data, { text: 'text', num: 1 }, 'visible row\'s data');
-    });
 
     // T583229
     QUnit.test('The same page should not load when scrolling in virtual mode', function(assert) {
@@ -6428,90 +6157,6 @@ QUnit.module('Virtual row rendering', baseModuleConfig, () => {
         assert.ok($(dataGrid.element()).find('.dx-virtual-row').last().position().top >= dataGrid.getScrollable().scrollTop(), 'second virtual row is not in viewport');
     });
 
-    function createRemoteDataSourceWithGroupPaging(arrayStore, key) {
-        return {
-            key,
-            load: function(options) {
-                const d = $.Deferred();
-                setTimeout(function() {
-                    const result = {};
-                    arrayStore.load(options).done(function(data) {
-                        result.data = data;
-
-                        if(options.group) {
-                            data.forEach(item => {
-                                item.count = item.items.length;
-                                item.items = null;
-                            });
-                        }
-                    });
-                    if(options.requireGroupCount) {
-                        arrayStore.load({ filter: options.filter, group: options.group }).done(function(groupedData) {
-                            result.groupCount = groupedData.length;
-                        });
-                    }
-                    if(options.requireTotalCount) {
-                        arrayStore.totalCount(options).done(function(totalCount) {
-                            result.totalCount = totalCount;
-                        });
-                    }
-
-                    d.resolve(result);
-                }, 10);
-
-                return d;
-            }
-        };
-    }
-
-    // T716207
-    QUnit.test('Filtering should works correctly if groupPaging is enabled and group is expanded', function(assert) {
-    // arrange
-
-        const arrayStore = new ArrayStore([
-            { id: 1, group: 'group', type: 1 },
-            { id: 2, group: 'group', type: 1 },
-            { id: 3, group: 'group', type: 1 },
-            { id: 4, group: 'group', type: 2 }
-        ]);
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            dataSource: createRemoteDataSourceWithGroupPaging(arrayStore, 'id'),
-            remoteOperations: { groupPaging: true },
-            height: 400,
-            filterSyncEnabled: true,
-            loadingTimeout: undefined,
-            scrolling: {
-                mode: 'virtual'
-            },
-            grouping: {
-                autoExpandAll: false
-            },
-            paging: {
-                pageSize: 2
-            },
-            columns: [{
-                dataField: 'group',
-                groupIndex: 0
-            }, {
-                dataField: 'id'
-            }, {
-                dataField: 'type'
-            }]
-        }).dxDataGrid('instance');
-        this.clock.tick(100);
-
-        dataGrid.expandRow(['group']);
-        this.clock.tick(100);
-
-        // act
-        dataGrid.columnOption('type', 'filterValue', 1);
-        this.clock.tick(100);
-
-        // assert
-        assert.notOk(dataGrid.getDataSource().isLoading(), 'not loading');
-        assert.equal(dataGrid.getVisibleRows().length, 4, 'visible row count is correct');
-    });
-
     // T641931
     QUnit.test('Infinite scrolling should works correctly', function(assert) {
     // arrange, act
@@ -6895,110 +6540,9 @@ QUnit.module('Rendered on server', baseModuleConfig, () => {
         assert.equal(dataGrid.getVisibleRows().length, 1, 'visible rows are filtered');
         assert.equal(dataGrid.$element().find('.dx-data-row').length, 1, 'filtered data rows are rendered');
     });
-
-    // T655083
-    QUnit.test('Virtual rows should be hidden after filtering if cellTemplate is asynchronous', function(assert) {
-        const items = [];
-        for(let i = 0; i < 100; i++) {
-            items.push({ id: i + 1 });
-        }
-
-        const dataGrid = createDataGrid({
-            height: 500,
-            dataSource: items,
-            scrolling: {
-                mode: 'virtual'
-            },
-            columns: [{
-                dataField: 'id',
-                cellTemplate: function($container, options) {
-                    setTimeout(function() {
-                        $('<div>').text(options.text).appendTo($container);
-                    });
-                }
-            }]
-        });
-
-        // act
-        this.clock.tick(300);
-
-        // assert
-        assert.equal(dataGrid.$element().find('.dx-virtual-row').length, 1, '1 virtual rows');
-
-        // act
-        dataGrid.columnOption('id', 'filterValue', '99');
-        this.clock.tick(300);
-
-        // assert
-        assert.equal(dataGrid.getVisibleRows().length, 1, '1 visible row');
-        assert.equal(dataGrid.$element().find('.dx-virtual-row').length, 0, 'no virtual rows');
-    });
-
-    // T621703
-    QUnit.testInActiveWindow('Edit cell on onContentReady', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            dataSource: [{ firstName: 'Andrey', lastName: 'Prohorov' }],
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            onContentReady: function(e) {
-            // act
-                e.component.editCell(0, 1);
-            }
-        });
-
-        this.clock.tick();
-
-        // assert
-        const $cellElement = $(dataGrid.getCellElement(0, 1));
-        assert.ok($cellElement.hasClass('dx-editor-cell'), 'cell has editor');
-        assert.ok($cellElement.find('.dx-texteditor-input').is(':focus'), 'cell editor is focused');
-    });
 });
 
 QUnit.module('Async render', baseModuleConfig, () => {
-
-    QUnit.test('filterRow, command column and showEditorAlways column should render asynchronously if renderAsync is true', function(assert) {
-        let cellPreparedCells = [];
-
-        // act
-        createDataGrid({
-            dataSource: [{ id: 1, boolean: true }],
-            loadingTimeout: undefined,
-            renderAsync: true,
-            filterRow: {
-                visible: true
-            },
-            selection: {
-                mode: 'multiple'
-            },
-            onCellPrepared: function(e) {
-                cellPreparedCells.push(e.rowType + '-' + (e.column.command || e.column.dataField));
-            }
-        });
-
-
-        // assert
-        assert.deepEqual(cellPreparedCells, [
-            'header-id', 'header-boolean',
-            'filter-select', 'header-select',
-            'data-id'
-        ], 'synchronous cellPrepared calls');
-
-        // act
-        cellPreparedCells = [];
-        this.clock.tick();
-
-        // assert
-        assert.deepEqual(cellPreparedCells, [
-            'filter-id', 'filter-boolean', // filter row is async
-            'data-select', // command column is async
-            'data-boolean' // showEditorAlways column is async
-        ], 'asynchronous cellPrepared calls');
-    });
-
     QUnit.test('Template in columns.buttons should render asynchronously if column renderAsync is true (T876950)', function(assert) {
         let buttonTemplateCallCount = 0;
         const dataGrid = createDataGrid({
@@ -7570,83 +7114,6 @@ QUnit.module('Assign options', baseModuleConfig, () => {
         assert.equal(headerPanel._getToolbarOptions.callCount, 6, 'Toolbar items are updated after searchPanel options change');
     });
 
-    QUnit.test('search editor have not been recreated when search text is changed', function(assert) {
-    // arrange, act
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            remoteOperations: { filtering: true, sorting: true, paging: true },
-            searchPanel: {
-                visible: true,
-            },
-            dataSource: {
-                store: {
-                    type: 'array',
-                    data: [{ a: 1, b: 2 }, { a: 2, b: 1 }]
-                }
-            }
-        });
-        const searchEditor = $(dataGrid.$element()).find('.dx-datagrid-search-panel').dxTextBox('instance');
-        // act
-        dataGrid.option('searchPanel.text', '123');
-        // assert
-        assert.strictEqual(searchEditor.option('value'), '123');
-    });
-
-    // T744851
-    QUnit.test('search editor value should be changed when search text is changed if grid is rendered in toolbar', function(assert) {
-    // arrange, act
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            onToolbarPreparing: function(e) {
-                e.toolbarOptions.items.unshift({
-                    location: 'before',
-                    template: function() {
-                        return $('<div>').dxDataGrid({
-                            loadingTimeout: undefined,
-                            searchPanel: {
-                                visible: true
-                            }
-                        });
-                    }
-                });
-            },
-            searchPanel: {
-                visible: true,
-            },
-            dataSource: [{ a: 1, b: 2 }, { a: 2, b: 1 }]
-        });
-        const $searchEditors = $(dataGrid.$element()).find('.dx-datagrid-search-panel');
-
-        // act
-        dataGrid.option('searchPanel.text', '123');
-
-        // assert
-        assert.strictEqual($searchEditors.eq(0).dxTextBox('instance').option('value'), '', 'first search editor is not changed');
-        assert.strictEqual($searchEditors.eq(1).dxTextBox('instance').option('value'), '123', 'second search editor is changed');
-    });
-
-    QUnit.test('search editor have not been recreated on typing', function(assert) {
-    // arrange, act
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            remoteOperations: { filtering: true, sorting: true, paging: true },
-            searchPanel: {
-                visible: true,
-            },
-            dataSource: {
-                store: {
-                    type: 'array',
-                    data: [{ a: 1, b: 2 }, { a: 2, b: 1 }]
-                }
-            }
-        });
-        const searchEditor = $(dataGrid.$element()).find('.dx-datagrid-search-panel').dxTextBox('instance');
-        // act
-        searchEditor.option('value', '123');
-        // assert
-        assert.strictEqual(searchEditor, $(dataGrid.$element()).find('.dx-datagrid-search-panel').dxTextBox('instance'));
-    });
-
     QUnit.test('customizeColumns change', function(assert) {
     // arrange, act
         const dataGrid = createDataGrid({
@@ -7958,27 +7425,6 @@ QUnit.module('Assign options', baseModuleConfig, () => {
         // assert
         assert.ok($(dataGrid.$element()).find('.dx-datagrid-rowsview').height() > 0);
         assert.equal($(dataGrid.$element()).find('.dx-datagrid-bottom-load-panel').length, 0);
-    });
-
-    QUnit.test('filterRow.visible change after clearFilter', function(assert) {
-    // arrange, act
-        const dataGrid = createDataGrid({
-            dataSource: [{ a: 1111, b: 222 }]
-        });
-
-        this.clock.tick();
-
-        // act
-        dataGrid.clearFilter();
-        dataGrid.option('filterRow.visible', true);
-
-        this.clock.tick();
-
-        // assert
-        assert.equal($(dataGrid.$element()).find('.dx-datagrid-filter-row').length, 1, 'filter row is rendered');
-
-        assert.strictEqual(dataGrid.getView('columnHeadersView')._requireReady, false, 'columnHeadersView requireReady is false');
-        assert.strictEqual(dataGrid.getView('rowsView')._requireReady, false, 'rowsView requireReady is false');
     });
 
     // T210836
@@ -8630,33 +8076,6 @@ QUnit.module('Assign options', baseModuleConfig, () => {
 
         // assert
         assert.strictEqual(onOptionChanged.callCount, 1, 'onOptionChanged is called once');
-    });
-
-    // T837684
-    QUnit.test('There are no exceptions when changing a filterValue to an array and selectedFilterOperation to \'between\' for date column', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ field: new Date(1992, 7, 6) }, { field: new Date(1992, 7, 9) }],
-            filterRow: { visible: true },
-            columns: [{ dataField: 'field', dataType: 'date' }]
-        });
-
-        // assert
-        assert.strictEqual(dataGrid.getVisibleRows().length, 2, 'row count');
-
-        try {
-        // act
-            dataGrid.option('columns[0].filterValue', [new Date(1992, 7, 5), new Date(1992, 7, 7)]);
-            dataGrid.option('columns[0].selectedFilterOperation', 'between');
-
-            // assert
-            assert.ok(true, 'no exceptions');
-            assert.strictEqual(dataGrid.getVisibleRows().length, 1, 'row count');
-        } catch(e) {
-        // assert
-            assert.ok(false, 'exception');
-        }
     });
 });
 
@@ -10246,61 +9665,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.equal(dataGrid.columnOption(1, 'sortOrder'), 'asc', 'sortOrder was returned to default');
     });
 
-    // T818434
-    QUnit.test('State reset should reset filtering', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            columns: [{ dataField: 'field1' }, { dataField: 'field2' }],
-            filterPanel: { visible: true },
-            dataSource: [{ field1: 'test1', field2: 1 }, { field1: 'test2', field2: 2 }]
-        });
-        let filter;
-
-        this.clock.tick(0);
-
-        // act
-        filter = ['field1', '=', 'test1'];
-        dataGrid.option('filterValue', filter);
-
-        // assert
-        assert.deepEqual(dataGrid.option('filterValue'), filter, 'dataGrid\'s filter');
-
-        // act
-        dataGrid.state(null);
-        this.clock.tick(0);
-
-        // assert
-        assert.equal(dataGrid.option('filterValue'), undefined, 'dataGrid\'s filter');
-
-        // act
-        filter = ['field2', '=', 1];
-        dataGrid.option('filterValue', filter);
-
-        // assert
-        assert.deepEqual(dataGrid.option('filterValue'), filter, 'dataGrid\'s filter');
-
-        // act
-        dataGrid.state(null);
-        this.clock.tick(0);
-
-        // assert
-        assert.equal(dataGrid.option('filterValue'), undefined, 'dataGrid\'s filter');
-
-        // act
-        filter = [['field1', '=', 'test1'], 'and', ['field2', '=', 1]];
-        dataGrid.option('filterValue', filter);
-
-        // assert
-        assert.deepEqual(dataGrid.option('filterValue'), filter, 'dataGrid\'s filter');
-
-        // act
-        dataGrid.state(null);
-        this.clock.tick(0);
-
-        // assert
-        assert.equal(dataGrid.option('filterValue'), undefined, 'dataGrid\'s filter');
-    });
-
     // T817555
     QUnit.test('State reset should return default grouping and sorting after multiple changes', function(assert) {
     // arrange
@@ -10327,31 +9691,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
 
         assert.equal(dataGrid.columnOption(0, 'sortOrder'), undefined, 'sortOrder was returned to default');
         assert.equal(dataGrid.columnOption(1, 'sortOrder'), 'asc', 'sortOrder was returned to default');
-    });
-
-    // T832870
-    QUnit.test('onEditorPreparing event should not be fired twice for each column if state storing and filter row are used', function(assert) {
-        let onEditorPreparingCallCount = 0;
-
-        createDataGrid({
-            columns: [{ dataField: 'field1' }, { dataField: 'field2' }],
-            dataSource: [{ field1: 'data', field2: 'data2' }],
-            onEditorPreparing: function(e) {
-                onEditorPreparingCallCount++;
-            },
-            filterRow: {
-                visible: true
-            },
-            stateStoring: {
-                enabled: true
-            }
-        });
-
-        // act
-        this.clock.tick();
-
-        // assert
-        assert.equal(onEditorPreparingCallCount, 2, 'onEditorPreparing call count');
     });
 
     QUnit.test('Clear state when initial options is defined in dataSource', function(assert) {
@@ -10451,91 +9790,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         // assert
         assert.equal(dataGrid.pageIndex(), 0, 'pageIndex');
         assert.equal(dataGrid.pageSize(), 2, 'pageSize');
-    });
-
-    // T735143
-    QUnit.test('Apply state with paging and filtering if filterPanel is visible', function(assert) {
-        const dataGrid = createDataGrid({
-            columns: ['id'],
-            dataSource: [{ id: 1 }, { id: 2 }, { id: 3 }, { id: 4 }],
-            paging: {
-                pageSize: 2,
-                pageIndex: 1
-            },
-            filterPanel: {
-                visible: true
-            },
-            headerFilter: {
-                visible: true
-            },
-            filterRow: {
-                visible: true
-            }
-        });
-
-        this.clock.tick();
-
-        // act
-        dataGrid.state({ pageIndex: 1, pageSize: 2, filterValue: ['id', '<>', 1] });
-        this.clock.tick();
-
-        // assert
-        assert.equal(dataGrid.pageIndex(), 1, 'pageIndex is applied');
-        assert.equal(dataGrid.getVisibleRows().length, 1, 'rows are filtered');
-    });
-
-    // T414555
-    QUnit.test('Apply state when search text and grouping are changed', function(assert) {
-        let loadingCount = 0;
-        const dataGrid = createDataGrid({
-            columns: [
-                'ID',
-                { dataField: 'Terms', groupIndex: 0 },
-                'Employee'
-            ],
-            dataSource: {
-                store: {
-                    type: 'array',
-                    data: [{
-                        'ID': 47,
-                        'Terms': '30 Days',
-                        'Employee': 'Clark Morgan'
-                    }],
-                    onLoading: function() {
-                        loadingCount++;
-                    }
-                }
-            },
-            stateStoring: {
-                ignoreColumnOptionNames: []
-            }
-        });
-
-        this.clock.tick();
-
-        assert.equal(dataGrid.columnOption('groupIndex:0').dataField, 'Terms', 'grouped column exists');
-
-        loadingCount = 0;
-
-        const strState = {
-            'columns': [
-                { 'visibleIndex': 0, 'dataField': 'ID', 'dataType': 'number', 'visible': true },
-                { 'visibleIndex': 1, 'dataField': 'Terms', 'dataType': 'string', 'visible': true, 'sortOrder': 'asc', 'sortIndex': 0 },
-                { 'visibleIndex': 2, 'dataField': 'Employee', 'dataType': 'string', 'visible': true }],
-            'searchText': 'A',
-            'pageIndex': 0,
-            'pageSize': 0,
-            'allowedPageSizes': []
-        };
-
-        // act
-        dataGrid.state(strState);
-        this.clock.tick();
-
-        // assert
-        assert.ok(!dataGrid.columnOption('groupIndex:0'), 'no grouped columns');
-        assert.equal(dataGrid.option('searchPanel.text'), 'A', 'search panel text is applied');
-        assert.equal(loadingCount, 1, 'loading count');
     });
 
     // T296786
@@ -11533,20 +10787,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.strictEqual($(dataGrid.getCellElement(1, 1)).text(), 'updated', 'second cell value is updated');
     });
 
-    // T443177
-    QUnit.test('Show searchPanel via option method', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({});
-
-        // act
-        dataGrid.option('searchPanel.visible', true);
-
-        // assert
-        const $headerPanelElement = $($(dataGrid.$element()).find('.dx-datagrid-header-panel'));
-        assert.ok($headerPanelElement.length, 'has headerPanel');
-        assert.ok($headerPanelElement.find('.dx-datagrid-search-panel').length, 'has searchPanel');
-    });
-
     // T548906
     QUnit.test('Change page index when virtual scrolling is enabled', function(assert) {
     // arrange
@@ -11574,40 +10814,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
 
         // assert
         assert.equal(dataGrid.pageIndex(), 3, 'page index');
-    });
-
-    // T548906
-    QUnit.test('Filtering on load when virtual scrolling', function(assert) {
-    // arrange
-        const generateDataSource = function(count) {
-            const result = []; let i;
-            for(i = 0; i < count; ++i) {
-                result.push({ firstName: 'name_' + i, lastName: 'lastName_' + i });
-            }
-            return result;
-        };
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            height: 50,
-            dataSource: generateDataSource(10),
-            scrolling: {
-                mode: 'virtual'
-            },
-            paging: {
-                pageSize: 2
-            },
-            columns: [
-                { dataField: 'firstName', filterValue: 'name_5' },
-                'lastName'
-            ]
-        });
-
-        const items = dataGrid.getDataSource().items();
-
-        // assert
-        assert.equal(items.length, 1, '1 item in dataSource');
-        assert.equal(items[0].firstName, 'name_5', 'filtered row \'firstName\' field value');
-        assert.equal(items[0].lastName, 'lastName_5', 'filtered row \'lastName\' field value');
     });
 
     QUnit.test('DataGrid should not paginate to the already loaded page if it is not in the viewport and it\'s row was focused (T726994)', function(assert) {
@@ -11643,41 +10849,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
 
         // assert
         assert.deepEqual(visibleRow0.key, dataGrid.getController('data').getVisibleRows()[0].key, 'Compare first visible row');
-    });
-
-    // T731090
-    QUnit.test('DataGrid should hide load panel after filtering to no data if focused row is enabled', function(assert) {
-    // arrange
-        const generateDataSource = function(count) {
-            const result = [];
-            for(let i = 1; i <= count; ++i) {
-                result.push({ id: i });
-            }
-            return result;
-        };
-        const dataGrid = createDataGrid({
-            height: 100,
-            dataSource: generateDataSource(100),
-            keyExpr: 'id',
-            focusedRowEnabled: true,
-            focusedRowKey: 1,
-            scrolling: {
-                mode: 'virtual'
-            }
-        });
-
-        this.clock.tick();
-
-        dataGrid.pageIndex(5);
-        this.clock.tick();
-
-        // act
-        dataGrid.filter(['id', '=', 666]);
-        this.clock.tick();
-
-        // assert
-        assert.strictEqual(dataGrid.getVisibleRows().length, 0, 'no rows');
-        assert.strictEqual(dataGrid.getController('data').isLoading(), false, 'no loading');
     });
 
     // T558189
@@ -11824,48 +10995,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
             'rowExpanded rowCount: 2, groupRow: true, dataRow: true',
             'done rowCount: 2, groupRow: true, dataRow: true'
         ]);
-    });
-
-    QUnit.testInActiveWindow('Filter row editor should have focus after _synchronizeColumns (T638737)', function(assert) {
-        $('#qunit-fixture').css('position', 'static');
-        // arrange, act
-        const dataGrid = createDataGrid({
-            filterRow: { visible: true },
-            editing: { allowAdding: true },
-            columns: [
-                { dataField: 'field1' },
-                { dataField: 'field2' }
-            ],
-            dataSource: [{ field1: 1, field2: 2 }, { field1: 3, field2: 4 }]
-        });
-
-        this.clock.tick();
-
-        const $input = $(dataGrid.$element()).find('.dx-editor-cell').first().find('.dx-texteditor-input');
-        $input.focus().val('1').trigger('change');
-
-        const selectionRangeArgs = [];
-
-        const oldSetSelectionRange = gridCoreUtils.setSelectionRange;
-        gridCoreUtils.setSelectionRange = function(element, range) {
-            oldSetSelectionRange.apply(this, arguments);
-            selectionRangeArgs.push([element, range]);
-        };
-
-        this.clock.tick();
-
-        gridCoreUtils.setSelectionRange = oldSetSelectionRange;
-
-        // assert
-        const $focusedInput = dataGrid.$element().find('.dx-editor-cell .dx-texteditor-input:focus');
-        assert.equal(dataGrid.getVisibleRows().length, 1, 'filter was applied');
-        assert.ok($focusedInput.length, 'filter cell has focus after filter applyed');
-        // T662207
-        if(devices.real().deviceType === 'desktop') {
-            assert.deepEqual(selectionRangeArgs, [[$focusedInput.get(0), { selectionStart: 1, selectionEnd: 1 }]], 'setSelectionRange args');
-        }
-
-        $('#qunit-fixture').css('position', '');
     });
 });
 
@@ -13177,72 +12306,6 @@ QUnit.module('Column Resizing', baseModuleConfig, () => {
         assert.strictEqual(instance.getVisibleRows().length, 8, 'row count');
         // T835869
         assert.strictEqual(loadingSpy.callCount, 1, 'data is loaded once');
-    });
-
-    // T596274
-    QUnit.testInActiveWindow('Resize a column with the \'between\' filter should not throw an exception', function(assert) {
-        // arrange
-        let $filterRangeContent;
-
-        fx.off = true;
-
-        try {
-            const dataGrid = $('#dataGrid').dxDataGrid({
-                width: 200,
-                allowColumnResizing: true,
-                loadingTimeout: undefined,
-                filterRow: {
-                    visible: true
-                },
-                dataSource: [{ name: 'Bob', age: 16 }],
-                columns: [
-                    { dataField: 'name', width: 100 },
-                    { dataField: 'age', width: 100, selectedFilterOperation: 'between' }
-                ]
-            });
-            const instance = dataGrid.dxDataGrid('instance');
-
-            $filterRangeContent = $('#dataGrid').find('.dx-datagrid-filter-row').find('.dx-filter-range-content').first();
-            $filterRangeContent.focus();
-            this.clock.tick();
-
-            // assert
-            assert.strictEqual($('.dx-overlay-wrapper.dx-datagrid-filter-range-overlay').length, 1, 'has overlay wrapper');
-
-            // act
-            const resizeController = instance.getController('columnsResizer');
-            resizeController._startResizing({
-                event: {
-                    data: resizeController,
-                    type: 'touchstart',
-                    pageX: -9900,
-                    pageY: -9990,
-                    preventDefault: function() {},
-                    stopPropagation: function() {}
-                }
-            });
-            resizeController._moveSeparator({
-                event: {
-                    data: resizeController,
-                    pageX: -9850,
-                    preventDefault: commonUtils.noop
-                }
-            });
-            resizeController._endResizing({
-                event: {
-                    data: resizeController
-                }
-            });
-
-            // assert
-            assert.strictEqual(instance.columnOption(0, 'width'), 150);
-            assert.strictEqual(instance.columnOption(1, 'width'), 50);
-        } catch(e) {
-            // assert
-            assert.ok(false, 'exception');
-        } finally {
-            fx.off = false;
-        }
     });
 
     // T527538
